Benchmarking for Envoy (#2)
Bencharking is being done by running a top-level Python script. The script creates a VM in Google Cloud and performs benchmarking inside the VM. In the end, it deletes the VM.

For now, it runs Nginx server and Envoy with a loopback configuration. h2load is used to generate load directly on Nginx server and through Envoy proxy. Result is saved as a JSON file.