Storing h2load results in a DB (#7)
store_data.py can be run to store data in a GCloud SQL instance. By this Python file, an DB instance will be created on GCloud. Then, a DB and a table on the DB will be created. A h2load JSON result file will be loaded, and corresponding data will be inserted into newly created table. The table will only be created if already not exists.