[salvo] Initial commit for benchmark abstraction framework (#73)
This is the initial 'official' commit for the Salvo tool. This aims to abstract the execution of nighthawk to benchmark a given envoy version.

See this issue for some background. The two design docs for this project are referenced here.

In this commit, salvo is placed into a separate directory of the envoy-perf repository, and is referenced from the main README.md.

Testing: Unit tests included, Address as many pylint3 issues as feasible
[#Issue] envoyproxy/envoy#11266

Signed-off-by: Alvin Baptiste <alvinsb@gmail.com>