Add the ability to specify additional benchmark images (#80)
Extend the `source` and `api` protos so that users can specify an arbitrary number of benchmark images.

The intent is to give the users the ability to have more than two points for benchmarks.  In the current mode,  the specified commit hash and one commit prior is tested.  With this change, the user can manually specify other points for which we test.

Signed-off-by: abaptiste <abaptiste@users.noreply.github.com>