Add Source Manager implementation and tests (#92)
This PR adds the implementation for the source manager object.  This source manager manipulates source trees and its purpose is to have a single interface through which all artifacts are compiled.

We also updated the `source.proto` to allow users to specify options that are supplied to bazel.  This allows us to build an Envoy binary that does not reject NightHawks attempts to enable profiling.

Signed-off-by: abaptiste <abaptiste@users.noreply.github.com>