[Salvo] Add documentation 'Measure Envoy's Performance Change with an A/B Testing' (#127)
Thank you for starting this @gyohuangxin. We have a good introduction of the user story and a set of commands to run Salvo.

What we should invest more time in is connecting the story to the conclusion. Salvo currently runs that one example test scenario (the one pytest target). That target wasn't developed with an exact goal in mind, it serves as an example of how to use the testing framework. It is unlikely that test will show useful results when it comes to comparing performance.

Additionally the users of Salvo are likely unaware of how to interpret the pytest and Nighthawk outputs. Which is another road block we should try to overcome.

What we could do is connect the dots from the user story down to the very end of interpreting the results. How would you feel about making the story very concrete. I.e. compare two commits where one is making a clear performance degradation in Envoy.  We can then develop a specific test that exercises this new hot path and show how to read the performance results. I.e. exactly prove how we can identify a performance degradation using Salvo and Nighthawk.

Happy to talk more on Slack or in a meeting to hash out the details.

Signed-off-by: Huang Xin <xin1.huang@intel.com>